Prefer `RB_OBJ_WRITE` over `RB_OBJ_WRITTEN`
This is nitpick territory, but when I'm auditing a C extension
for write barrier issues, what I do is that I go over all the
`VALUE` fields in strucs and for each for them I grep for `->field = `.

This makes it quick to ensure all places where the value is assigned
do trigger a write barrier.

With this workflow, `RB_OBJ_WRITTEN` end up causing false positive.