Mark the thread from the worker pool.
If we don't mark them, the thread object can be freed as soon as it ends.
It doesn't seem to be a problem when the worker pool just has 1 thread,
but if there are multiple threads then each call to `Thread#join` can
cause a GC step as it calls back into ruby.

If you do this:

```
class ::Thread
  alias_method :old_join, :join
  def join(timeout=nil)
    4.times { GC.start }
    old_join(timeout)
  end
end
```

and set the size pool to be > 1, it can crash.