Mark the thread from the worker pool. (#152)
* Mark the thread from the worker pool.

If we don't mark them, the thread object can be freed as soon as it ends.
It doesn't seem to be a problem when the worker pool just has 1 thread,
but if there are multiple threads then each call to `Thread#join` can
cause a GC step as it calls back into ruby.